+++ spring-cloud-dataflow-server-core/src/main/java/org/springframework/cloud/dataflow/server/controller/AppRegistryController.java
@@ -132,7 +132,7 @@ public class AppRegistryController implements ResourceLoaderAware {
 			// building new page.
 			// all this is done because we don't use a proper repository which would
 			// handle all these automatically.
-			int offset = 0;
+			long offset = 0;
 			int page = 0;
 			if (pageable.getOffset() <= to) {
 				offset = pageable.getOffset();
@@ -141,7 +141,7 @@ public class AppRegistryController implements ResourceLoaderAware {
 			else if (pageable.getOffset() + pageable.getPageSize() <= to) {
 				offset = pageable.getOffset();
 			}
-			pagedRegistrations = new PageImpl<>(appRegistrations.subList(offset, (int) to),
+			pagedRegistrations = new PageImpl<>(appRegistrations.subList((int) offset, (int) to),
 					new PageRequest(page, pageable.getPageSize()), appRegistrations.size());
 		}
